+++ src/os/filestore/JournalingObjectStore.h
@@ -19,6 +19,7 @@
 #include "Journal.h"
 #include "FileJournal.h"
 #include "common/RWLock.h"
+#include "osd/OpRequest.h"
 
 class JournalingObjectStore : public ObjectStore {
 protected:
